[update] DetailVCのアップデート
