+++ Drinto/DrinkMemory/Views/DetailDrinkViewController.swift
@@ -8,22 +8,37 @@
 import UIKit
 
 class DetailDrinkViewController: UIViewController {
+    @IBOutlet private weak var drinkNameLabel: UILabel!
+    @IBOutlet private weak var drinkCategoryLabel: UILabel!
+    @IBOutlet private weak var drinkImageView: UIImageView!
+    @IBOutlet private weak var chartView: ChartView!
+
+    var detailDrinkMemory = DrinkMemorySwiftModel()
+
+    private var drinkImageModel = DrinkImageModel()
+
+    private var drinkCategoryModel = DrinkCategoryModel()
+
+    private var drinkCategory: DrinkCategory?
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        drinkNameLabel.text = detailDrinkMemory.drinkName
+        drinkCategoryLabel.text = detailDrinkMemory.category
 
-        // Do any additional setup after loading the view.
+        drinkCategoryModel.drinkCategoryString = detailDrinkMemory.category
+
+        chartView.configure(valueInfo:
+                .init(
+                    value1: detailDrinkMemory.drinkPoint?[0] ?? 0,
+                    value2: detailDrinkMemory.drinkPoint?[1] ?? 0,
+                    value3: detailDrinkMemory.drinkPoint?[2] ?? 0,
+                    value4: detailDrinkMemory.drinkPoint?[3] ?? 0,
+                    value5: detailDrinkMemory.drinkPoint?[4] ?? 0,
+                    value6: detailDrinkMemory.drinkPoint?[5] ?? 0
+                ), drinkCategory: drinkCategoryModel.drinkCategoryStringToType()
+        )
+
+        drinkImageView.image = drinkImageModel.getImageData(imageUUID: detailDrinkMemory.uuidString)
     }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
 }
