+++ Drinto/Extension/SwinjectStoryboard+Setup.swift
@@ -1,0 +1,16 @@
+//
+//  SwinjectStoryboard+Setup.swift
+//  Drinto
+//
+//  Created by HiroakiSaito on 2023/05/05.
+//
+
+import Foundation
+import SwinjectStoryboard
+
+extension SwinjectStoryboard {
+
+//    @objc class setUp() {
+//
+//    }
+}
